add luci support

+++ src/sx1276.hpp
@@ -113,8 +113,8 @@ public:
   /// Predict time on air for a zero terminated payload.
   /// @param Payload text
   /// @return Time on air, seconds.
-  float PredictTimeOnAir(const char *payload) const;
-  float PredictTimeOnAir(const void *payload, unsigned len) const;
+  int PredictTimeOnAir(const char *payload) const;
+  int PredictTimeOnAir(const void *payload, unsigned len) const;
 
 private:
